Refactor venue and event views for HTMX support
Updated the venue and event views to add HTMX support for partial page updates. Introduced a `MultiViewContainer` component for consistent layout and added related routing in `app.py`. Removed the `EventsActions` function and included a new `NavMenu` for navigation.  Enhanced the naming of the layout with ids to make clear sections of the UI